Make the old tests in tests.rs work
