+++ rust/pact_mock_server/src/matching.rs
@@ -1,0 +1,96 @@
+use pact_matching::models::{Interaction, Request, PactSpecification};
+use pact_matching::Mismatch;
+use pact_matching::s;
+use serde_json::json;
+use itertools::Itertools;
+
+/// Enum to define a match result
+#[derive(Debug, Clone, PartialEq)]
+pub enum MatchResult {
+    /// Match result where the request was sucessfully matched
+    RequestMatch(Interaction),
+    /// Match result where there were a number of mismatches
+    RequestMismatch(Interaction, Vec<Mismatch>),
+    /// Match result where the request was not expected
+    RequestNotFound(Request),
+    /// Match result where an expected request was not received
+    MissingRequest(Interaction)
+}
+
+impl MatchResult {
+    /// Returns the match key for this mismatch
+    pub fn match_key(&self) -> String {
+        match self {
+            &MatchResult::RequestMatch(_) => s!("Request-Matched"),
+            &MatchResult::RequestMismatch(_, _) => s!("Request-Mismatch"),
+            &MatchResult::RequestNotFound(_) => s!("Unexpected-Request"),
+            &MatchResult::MissingRequest(_) => s!("Missing-Request")
+        }
+    }
+
+    /// Returns true if this match result is a `RequestMatch`
+    pub fn matched(&self) -> bool {
+        match self {
+            &MatchResult::RequestMatch(_) => true,
+            _ => false
+        }
+    }
+
+    /// Converts this match result to a `Value` struct
+    pub fn to_json(&self) -> serde_json::Value {
+        match self {
+            &MatchResult::RequestMatch(_) => json!({ s!("type") : s!("request-match")}),
+            &MatchResult::RequestMismatch(ref interaction, ref mismatches) => mismatches_to_json(&interaction.request, mismatches),
+            &MatchResult::RequestNotFound(ref req) => json!({
+                "type": json!("request-not-found"),
+                "method": json!(req.method),
+                "path": json!(req.path),
+                "request": req.to_json(&PactSpecification::V3)
+            }),
+            &MatchResult::MissingRequest(ref interaction) => json!({
+                "type": json!("missing-request"),
+                "method": json!(interaction.request.method),
+                "path": json!(interaction.request.path),
+                "request": interaction.request.to_json(&PactSpecification::V3)
+            })
+        }
+    }
+}
+
+fn mismatches_to_json(request: &Request, mismatches: &Vec<Mismatch>) -> serde_json::Value {
+    json!({
+        s!("type") : json!("request-mismatch"),
+        s!("method") : json!(request.method),
+        s!("path") : json!(request.path),
+        s!("mismatches") : mismatches.iter().map(|m| m.to_json()).collect::<serde_json::Value>()
+    })
+}
+
+fn method_or_path_mismatch(mismatches: &Vec<Mismatch>) -> bool {
+    mismatches.iter()
+        .map(|mismatch| mismatch.mismatch_type())
+        .any(|mismatch_type| mismatch_type == "MethodMismatch" || mismatch_type == "PathMismatch")
+}
+
+pub fn match_request(req: &Request, interactions: &Vec<Interaction>) -> MatchResult {
+    let match_results = interactions
+        .into_iter()
+        .map(|i| (i.clone(), pact_matching::match_request(i.request.clone(), req.clone())))
+        .sorted_by(|i1, i2| {
+            let list1 = i1.1.clone().into_iter().map(|m| m.mismatch_type()).unique().count();
+            let list2 = i2.1.clone().into_iter().map(|m| m.mismatch_type()).unique().count();
+            Ord::cmp(&list1, &list2)
+        });
+    match match_results.first() {
+        Some(res) => {
+            if res.1.is_empty() {
+                MatchResult::RequestMatch(res.0.clone())
+            } else if method_or_path_mismatch(&res.1) {
+                MatchResult::RequestNotFound(req.clone())
+            } else {
+                MatchResult::RequestMismatch(res.0.clone(), res.1.clone())
+            }
+        },
+        None => MatchResult::RequestNotFound(req.clone())
+    }
+}
